default async, crawl steam when sleeping
