+++ config/config.ini
@@ -25,8 +25,6 @@ url_cache_hour = 6
 force_crawl = False
 # 一次请求的超时重试次数
 retry_times = 3
-#是否使用异步
-crawl_steam_async = True
 
 # 基本参数设置
 [COMMON]
