+++ src/config/definitions.py
@@ -39,7 +39,6 @@ FREQUENCY_INTERVAL_HIGH = int(config_behavior['frequency_interval_high'])
 URL_CACHE_HOUR = int(config_behavior['url_cache_hour'])
 FORCE_CRAWL = config_behavior.getboolean('force_crawl')
 RETRY_TIMES = int(config_behavior['retry_times'])
-CRAWL_STEAM_ASYNC = config_behavior.getboolean('crawl_steam_async')
 
 # common
 config_common = configr['COMMON']
@@ -112,7 +111,6 @@ class config_export:
         self.URL_CACHE_HOUR = URL_CACHE_HOUR
         self.FORCE_CRAWL = FORCE_CRAWL
         self.RETRY_TIMES = RETRY_TIMES
-        self.CRAWL_STEAM_ASYNC = CRAWL_STEAM_ASYNC
         self.STEAM_SELL_TAX = STEAM_SELL_TAX
         self.CRAWL_MIN_PRICE_ITEM = CRAWL_MIN_PRICE_ITEM
         self.CRAWL_MAX_PRICE_ITEM = CRAWL_MAX_PRICE_ITEM
