+++ src/util/requester.py
@@ -43,7 +43,6 @@ def get_json_dict_raw(url, cookies = {}, proxy = False, times = 1, is_steam_requ
         log.error('Timeout for {} beyond the maximum({}) retry times. SKIP!'.format(url, config.RETRY_TIMES))
         return None
 
-    timer.sleep_awhile(is_steam_request)
     try:
         if proxy and config.PROXY != {}:
             return requests.get(url, headers = headers, cookies = cookies, timeout = 5,
